+++ src/01-document/categories.js
@@ -1,4 +1,4 @@
-const i18n = require('../data/i18n');
+const i18n = require('../_data/i18n');
 const cat_reg = new RegExp('\\[\\[:?(' + i18n.categories.join('|') + '):(.{2,60}?)]](w{0,10})', 'ig');
 const cat_remove_reg = new RegExp('^\\[\\[:?(' + i18n.categories.join('|') + '):', 'ig');
