+++ src/01-document/redirects.js
@@ -1,4 +1,4 @@
-const i18n = require('../data/i18n');
+const i18n = require('../_data/i18n');
 const parseLink = require('../04-sentence/links');
 //pulls target link out of redirect page
 const REDIRECT_REGEX = new RegExp('^[ \n\t]*?#(' + i18n.redirects.join('|') + ') *?(\\[\\[.{2,180}?\\]\\])', 'i');
