+++ src/02-section/heading.js
@@ -1,4 +1,4 @@
-const fns = require('../lib/helpers');
+const fns = require('../_lib/helpers');
 const parseSentence = require('../04-sentence/').oneSentence;
 const parseReferences = require('../reference/');
 const heading_reg = /^(={1,5})(.{1,200}?)={1,5}$/;
