+++ src/03-paragraph/toJson.js
@@ -1,4 +1,4 @@
-const setDefaults = require('../lib/setDefaults');
+const setDefaults = require('../_lib/setDefaults');
 
 const defaults = {
   sentences: true
