+++ src/04-sentence/index.js
@@ -1,10 +1,10 @@
-const helpers = require('../lib/helpers');
+const helpers = require('../_lib/helpers');
 const parseLinks = require('./links');
 const parseFmt = require('./formatting');
 const Sentence = require('./Sentence');
 // const templates = require('./templates');
 const sentenceParser = require('./sentence-parser');
-const i18n = require('../data/i18n');
+const i18n = require('../_data/i18n');
 const cat_reg = new RegExp(
   '\\[\\[:?(' + i18n.categories.join('|') + '):[^\\]\\]]{2,80}\\]\\]',
   'gi'
