+++ src/04-sentence/links.js
@@ -1,9 +1,9 @@
-// const helpers = require('../lib/helpers');
+// const helpers = require('../_lib/helpers');
 const parse_interwiki = require('./interwiki');
 const ignore_links = /^:?(category|catégorie|Kategorie|Categoría|Categoria|Categorie|Kategoria|تصنيف|image|file|image|fichier|datei|media):/i;
 const external_link = /\[(https?|news|ftp|mailto|gopher|irc)(:\/\/[^\]\| ]{4,1500})([\| ].*?)?\]/g;
 const link_reg = /\[\[(.{0,160}?)\]\]([a-z']+)?(\w{0,10})/gi; //allow dangling suffixes - "[[flanders]]'s"
-// const i18n = require('../data/i18n');
+// const i18n = require('../_data/i18n');
 // const isFile = new RegExp('(' + i18n.images.concat(i18n.files).join('|') + '):', 'i');
 
 const external_links = function(links, str) {
