+++ src/_data/abbreviations.js
