+++ src/_data/i18n.js
