+++ src/_lib/encode.js
