+++ src/_lib/setDefaults.js
