+++ src/_lib/smartReplace.js
